refactor(nodes): extract shared resolvePath utility, use in map
- Add utils/resolve-path.ts as canonical path resolver
- Supports dot notation, keyed array (contacts[0]), standalone
  index ([0].name), and empty paths
- map.ts now imports from shared utility
